repair string and thats all

+++ ft_printf_bonus.h
@@ -37,8 +37,8 @@ char	*ft_strjoin(char const *s1, char const *s2);
 int     ft_strlen(const char *str);
 void	ft_putnbr(long n, flg_stc *flg);
 void	ft_bzero(void *dest, size_t n);
-void	ft_nbr(unsigned long int n, flg_stc *flg);
-int ft_ulen(unsigned long int n);
+void	ft_nbr(unsigned int n, flg_stc *flg);
+int ft_ulen(unsigned int n);
 void ft_htoi(unsigned int num,flg_stc *flg, char *base);
 void ft_OX(char *str, flg_stc *flg);
 
@@ -51,7 +51,7 @@ int ft_string(char *str, flg_stc *flg);
 void ft_filler(char fill, int counter, flg_stc *flg);
 void ft_hex(unsigned long int n,flg_stc *flg, char *base);
 void ft_void_pointer(void *ptr, flg_stc *flg, char *base);
-void ft_utoi(unsigned long int num,flg_stc *flg);
+void ft_utoi(unsigned int num,flg_stc *flg);
 int ft_numlen(long int n);
